Redone how form is contained with container divs
Redid how the entire form body was contained. Instead of the left and right side being children of the container for the entire project, I've added another container inbetween to accurately determine the size of the container, to make it easier in the future to style it and make it responsive.